pyproject.toml
[tool.poetry]
name = "pt750"
version = "0.2.1"
description = "label printing tools for brother PT-P750W label maker"
authors = ["rpedde <ron@pedde.com>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "*"
treepoem = "*"
fastapi = {extras = ["all"], version = "*"}
Jinja2 = "*"
uvicorn = "*"
PyYAML = "*"
easysnmp = "*"
opencv-python = "^4.11.0.86"
httpx = "^0.28.1"
pydantic-settings = "^2.7.1"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
flake8-import-order = "*"
pytest = "*"
mypy = "*"
jsf = "*"
requests = "*"

[tool.poetry.scripts]
makelabel = "pt750.main:main"

[[tool.mypy.overrides]]
module = "PIL"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsf"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
